+++ sdk/eventhubs/azure-messaging-eventhubs/src/samples/java/com/azure/messaging/eventhubs/PublishEventsWithAzureIdentity.java
@@ -1,0 +1,105 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+package com.azure.messaging.eventhubs;
+
+import com.azure.core.credential.TokenCredential;
+import com.azure.identity.DefaultAzureCredentialBuilder;
+import reactor.core.Exceptions;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
+
+import java.time.Duration;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicReference;
+
+import static java.nio.charset.StandardCharsets.UTF_8;
+
+/**
+ * Sample demonstrates how to send an {@link EventDataBatch} to an Azure Event Hub using Azure Identity.
+ */
+public class PublishEventsWithAzureIdentity {
+    private static final Duration OPERATION_TIMEOUT = Duration.ofSeconds(30);
+
+    /**
+     * Main method to invoke this demo on how to send an {@link EventDataBatch} to an Azure Event Hub.
+     *
+     * @param args Unused arguments to the program.
+     */
+    public static void main(String[] args) {
+        Flux<EventData> telemetryEvents = Flux.just(
+            new EventData("Roast beef".getBytes(UTF_8)),
+            new EventData("Cheese".getBytes(UTF_8)),
+            new EventData("Tofu".getBytes(UTF_8)),
+            new EventData("Turkey".getBytes(UTF_8)));
+
+        // The default azure credential checks multiple locations for credentials and determines the best one to use.
+        // For the purpose of this sample, create a service principal and set the following environment variables.
+        // See https://docs.microsoft.com/azure/active-directory/develop/howto-create-service-principal-portal for
+        // information on how to create a service principal.
+        System.setProperty("AZURE_CLIENT_ID", "<<insert-service-principal-client-id>>");
+        System.setProperty("AZURE_CLIENT_ID", "<<insert-service-principal-client-application-secret>>");
+        System.setProperty("AZURE_TENANT_ID", "<<insert-service-principal-tenant-id>>");
+
+        // DefaultAzureCredentialBuilder exists inside the azure-identity package.
+        TokenCredential credential = new DefaultAzureCredentialBuilder()
+            .build();
+
+        // Create a producer.
+        // "<<fully-qualified-namespace>>" will look similar to "{your-namespace}.servicebus.windows.net"
+        // "<<event-hub-name>>" will be the name of the Event Hub instance you created inside the Event Hubs namespace.
+        EventHubProducerAsyncClient producer = new EventHubClientBuilder()
+            .credential(
+                "<<fully-qualified-namespace>>",
+                "<<event-hub-name>>",
+                credential)
+            .buildAsyncProducerClient();
+
+        final AtomicReference<EventDataBatch> currentBatch = new AtomicReference<>(
+            producer.createBatch().block());
+
+        // The sample Flux contains three events, but it could be an infinite stream of telemetry events.
+        telemetryEvents.flatMap(event -> {
+            final EventDataBatch batch = currentBatch.get();
+            if (batch.tryAdd(event)) {
+                return Mono.empty();
+            }
+
+            // The batch is full, so we create a new batch and send the batch. Mono.when completes when both operations
+            // have completed.
+            return Mono.when(
+                producer.send(batch),
+                producer.createBatch().map(newBatch -> {
+                    currentBatch.set(newBatch);
+
+                    // Add that event that we couldn't before.
+                    if (!newBatch.tryAdd(event)) {
+                        throw Exceptions.propagate(new IllegalArgumentException(String.format(
+                            "Event is too large for an empty batch. Max size: %s. Event: %s",
+                            newBatch.getMaxSizeInBytes(), event.getBodyAsString())));
+                    }
+
+                    return newBatch;
+                }));
+        }).then()
+            .doFinally(signal -> {
+                final EventDataBatch batch = currentBatch.getAndSet(null);
+                if (batch != null) {
+                    producer.send(batch).block(OPERATION_TIMEOUT);
+                }
+            })
+            .subscribe(unused -> System.out.println("Complete"),
+                error -> System.out.println("Error sending events: " + error),
+                () -> System.out.println("Completed sending events."));
+
+        // The .subscribe() creation and assignment is not a blocking call. For the purpose of this example, we sleep
+        // the thread so the program does not end before the send operation is complete. Using .block() instead of
+        // .subscribe() will turn this into a synchronous call.
+        try {
+            TimeUnit.SECONDS.sleep(5);
+        } catch (InterruptedException ignored) {
+        } finally {
+            // Disposing of our producer.
+            producer.close();
+        }
+    }
+}