+++ sdk/eventhubs/azure-messaging-eventhubs/src/test/java/com/azure/messaging/eventhubs/implementation/ReactorConnectionIntegrationTest.java
@@ -6,7 +6,7 @@ package com.azure.messaging.eventhubs.implementation;
 import com.azure.core.amqp.AmqpTransportType;
 import com.azure.core.amqp.ProxyOptions;
 import com.azure.core.amqp.implementation.AzureTokenManagerProvider;
-import com.azure.core.amqp.implementation.CBSAuthorizationType;
+import com.azure.core.amqp.implementation.CbsAuthorizationType;
 import com.azure.core.amqp.implementation.ClaimsBasedSecurityChannel;
 import com.azure.core.amqp.implementation.ConnectionOptions;
 import com.azure.core.amqp.implementation.ConnectionStringProperties;
@@ -27,7 +27,7 @@ import reactor.test.StepVerifier;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 
-import static com.azure.core.amqp.implementation.CBSAuthorizationType.SHARED_ACCESS_SIGNATURE;
+import static com.azure.core.amqp.implementation.CbsAuthorizationType.SHARED_ACCESS_SIGNATURE;
 
 public class ReactorConnectionIntegrationTest extends IntegrationTestBase {
 
@@ -80,7 +80,7 @@ public class ReactorConnectionIntegrationTest extends IntegrationTestBase {
     public void getCbsNodeAuthorize() {
         // Arrange
         final AzureTokenManagerProvider provider = new AzureTokenManagerProvider(
-            CBSAuthorizationType.SHARED_ACCESS_SIGNATURE,
+            CbsAuthorizationType.SHARED_ACCESS_SIGNATURE,
             getConnectionStringProperties().getEndpoint().getHost(),
             ClientConstants.AZURE_ACTIVE_DIRECTORY_SCOPE);