+++ sdk/eventhubs/azure-messaging-eventhubs-checkpointstore-blob/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Release History
+## 1.0.0-beta.4 (2019-12-02)
+- Artifact name changed from `preview` to `beta`.
+- `BlobCheckpointStore` implementation updated to match changes in `CheckpointStore` interface.
+
 ## 1.0.0-preview.3 (2019-11-01)
 - Renamed `BlobPartitionManager` to `BlobEventProcessorStore`
 - Added fully qualified namespace to list ownership API.